Pick up EmbeddedResouce items again
Commit 54f3ba64 set `EnableDefaultEmbeddedResourceItems` to `false`, meaning we no longer automatically pick up all .resx files under the project root as `EmbeddedResource` items. We *do* want to embed WizardResource.resx, however, and it already had an explicit item in the project file. However, the item specified the file with `Update` rather than `Include`, meaning we were going to update an existing item with the given name with new metadata rather than define a new item. This made sense when we were automatically picking up the file, but once we stopped doing that there was nothing to update.

The end result was that the resources in WizardResource.resx weren't being embedded during the build, we would fail to find them at runtime, and trying to add a .editorconfig file from the item template would result in an exception.

The fix here is to use `Include` instead of `Update`.