Properly determine the path for a new .editorconfig
Currently, if you select an existing _file_ in Solution Explorer and use Ctrl+Shift+A to try to add .editorconfig file, you'll see a dialog pop up with an odd message:

> Could not find a part of the path `C:\users\Tom\source\repos\ConsoleApp999\Program.cs\.editorconfig'.

The problem here is that the template wizard for .editorconfig files effectively assumes that if an item is selected in Solution Explorer (as opposed to a project or the top-level solution item) then that item must be a folder. It then treats the full path of that item as a path to a directory, rather than a path.

The fix here is to explicitly check if the selected item is a physical file or folder. If it is a folder we can continue to treat the full path as a directory; if it is a file then we need to trim off the file name.